+++ src/Models/EventLog.php
@@ -176,6 +176,7 @@ class EventLog extends Model
 
     /**
      * @return array<string, string> */
+    #[\Override]
     protected function casts(): array
     {
         return [