+++ src/Models/Subscription.php
@@ -214,6 +214,7 @@ class Subscription extends Model
     /**
      * @return array<string, string>
      */
+    #[\Override]
     protected function casts(): array
     {
         return [